requirements.txt
konlpy==0.6.0
loguru==0.7.2
matplotlib==3.9.2
numpy==1.26.4
Pillow==10.4.0
pydantic==2.9.2
pypdf==5.1.0
wordcloud==1.9.4
